+++ contracts/interfaces/IFactory.sol
@@ -32,12 +32,14 @@ interface IFactory {
     function pendingOwner() external view returns (address);
 
     /// @dev Return the fee per second earned by liquidity providers.
+    /// @dev Must be downcasted to uint16.
     /// @return The fee following UQ0.40 format.
-    function fee() external view returns (uint16);
+    function fee() external view returns (uint256);
 
     /// @dev Return the protocol fee per second earned by the owner.
+    /// @dev Must be downcasted to uint16.
     /// @return The protocol fee per second following UQ0.40 format.
-    function protocolFee() external view returns (uint16);
+    function protocolFee() external view returns (uint256);
 
     /// @dev Returns the address of a deployed pair.
     /// @param asset The address of the ERC20 being lent and borrowed.
